chore: move all containers to ItineraryPageContainer level
